+++ app/src/main/java/org/ahomewithin/ahomewithin/activities/MainActivity.java
@@ -16,6 +16,9 @@ import android.view.MenuItem;
 import android.view.View;
 
 import org.ahomewithin.ahomewithin.R;
+import org.ahomewithin.ahomewithin.fragments.HomeFragment;
+import org.ahomewithin.ahomewithin.fragments.LearnMoreFragment;
+import org.ahomewithin.ahomewithin.fragments.ServicesFragment;
 import org.ahomewithin.ahomewithin.fragments.StreamPagerFragment;
 
 import java.io.IOException;
